# ARROW-10682: [Rust] Improve sort kernel performance by enabling inlining of is_valid calls