Benchmark results on an aws t3 instance (cascadelake):

```
$ cargo bench --bench sort_kernel
sort 2^10               time:   [168.90 us 168.98 us 169.08 us]                      
                        change: [-59.182% -59.041% -58.824%] (p = 0.00 < 0.05)
                        Performance has improved.

sort 2^12               time:   [839.88 us 840.15 us 840.46 us]                      
                        change: [-57.095% -57.040% -56.947%] (p = 0.00 < 0.05)
                        Performance has improved.

sort nulls 2^10         time:   [148.69 us 148.75 us 148.82 us]                            
                        change: [-66.823% -66.727% -66.633%] (p = 0.00 < 0.05)
                        Performance has improved.

sort nulls 2^12         time:   [721.79 us 721.95 us 722.12 us]                            
                        change: [-65.966% -65.920% -65.845%] (p = 0.00 < 0.05)
                        Performance has improved.
```