@jhorstmann 2020-11-21
@jorgecarleitao I see you're currently also working on sorting, this might be of interest to you.